requirements.txt
streamlit
requests
python-dotenv
pandas
plotly
openai
